SQLite step wrapper instead of print, header file
New type step_t, which is a typedef'd sqlite3_stmt, to hide inner
workings of the API from the outside. This is now returned from
list_tags and filter_tag. The results from the query is aquired with the
new step_result function.

The header file ftag.h is introduced to show the public API.

Extra: whitespace errors are removed.